feat(tui): implement 7 missing vim normal-mode operations
Add D, C, X, ~, o, O, r to the vim normal-mode command set.

- D: delete cursor to end of line (d$), yanks to register
- C: delete cursor to end of line then enter Insert (c$)
- X: delete char before cursor (dh), no-op at column 0
- ~: toggle case of char at cursor and advance one position
- o: open blank line after current line, enter Insert
- O: open blank line before current line, enter Insert
- r: replace single char at cursor with next keypress;
     Esc/non-char silently cancels via pending_replace flag

Each mutation pushes an undo snapshot and clears the redo stack.
All new VimCommand variants are wired into handle_extra_command,
extra_normal_command, and the keybinding description table in
wonder-of-u-cli. Nine new unit tests cover all seven operations.

+++ crates/wonder-of-u-core/src/tips.rs
@@ -173,7 +173,7 @@ pub struct TipContext {
 /// - it has been shown within its `cooldown_sessions` window.
 #[must_use]
 pub fn is_tip_eligible(tip: &Tip, ctx: &TipContext, sessions_since_shown: u64) -> bool {
-    if sessions_since_shown < tip.cooldown_sessions as u64 {
+    if sessions_since_shown < tip.cooldown_sessions {
         return false;
     }
